Bug #607 make `vrf_name` immutable in `apstra_datacenter_routing_zone` resource

+++ apstra/test_utils/vni_pool.go
@@ -2,13 +2,14 @@ package testutils
 
 import (
 	"context"
+	"testing"
+
 	"github.com/Juniper/apstra-go-sdk/apstra"
 	"github.com/hashicorp/terraform-plugin-testing/helper/acctest"
 	"github.com/stretchr/testify/require"
-	"testing"
 )
 
-func VniPool(t testing.TB, ctx context.Context, first, last uint32) *apstra.VniPool {
+func VniPool(t testing.TB, ctx context.Context, first, last uint32, cleanup bool) *apstra.VniPool {
 	t.Helper()
 
 	client := GetTestClient(t, ctx)
@@ -23,7 +24,10 @@ func VniPool(t testing.TB, ctx context.Context, first, last uint32) *apstra.VniP
 		},
 	})
 	require.NoError(t, err)
-	t.Cleanup(func() { require.NoError(t, client.DeleteVniPool(ctx, id)) })
+
+	if cleanup {
+		t.Cleanup(func() { require.NoError(t, client.DeleteVniPool(ctx, id)) })
+	}
 
 	pool, err := client.GetVniPool(ctx, id)
 	require.NoError(t, err)
